refactor(ad-form): reorganize code
