+++ components/AdForm/AdOptions/AdSense.js
@@ -1,17 +1,24 @@
 /* eslint-disable jsx-a11y/label-has-for */
 import React from 'react';
 import PropTypes from 'prop-types';
-import TextInput from '../TextInput';
-import { required, mustBeNumber, composeValidators } from '../../validators';
+import TextInput from '../../TextInput';
+import { required, mustBeNumber, composeValidators } from '../../../validators';
+import { toNumber } from '../formats';
 
 const requireNumber = composeValidators(required, mustBeNumber);
 
-const DoubleClick = ({ member }) => (
+const AdSense = ({ member }) => (
   <>
+    <TextInput
+      name={`${member}.client`}
+      label="client"
+      placeholder="ca-pub-1234123412341234"
+      validate={required}
+    />
     <TextInput
       name={`${member}.slot`}
       label="slot"
-      placeholder="/12345678/ad-roba_1"
+      placeholder="1234567890"
       validate={required}
     />
     <TextInput
@@ -19,20 +26,25 @@ const DoubleClick = ({ member }) => (
       label="width"
       placeholder="300"
       validate={requireNumber}
-      parse={input => Number(input)}
+      parse={toNumber}
     />
     <TextInput
       name={`${member}.height`}
       label="height"
       placeholder="250"
       validate={requireNumber}
-      parse={input => Number(input)}
+      parse={toNumber}
+    />
+    <TextInput
+      name={`${member}.format`}
+      label="format"
+      placeholder="rectangle"
     />
   </>
 );
 
-DoubleClick.propTypes = {
+AdSense.propTypes = {
   member: PropTypes.string.isRequired,
 };
 
-export default DoubleClick;
+export default AdSense;
