+++ components/AdForm/AdOptions/DoubleClick.js
@@ -1,23 +1,18 @@
 /* eslint-disable jsx-a11y/label-has-for */
 import React from 'react';
 import PropTypes from 'prop-types';
-import TextInput from '../TextInput';
-import { required, mustBeNumber, composeValidators } from '../../validators';
+import TextInput from '../../TextInput';
+import { required, mustBeNumber, composeValidators } from '../../../validators';
+import { toNumber } from '../formats';
 
 const requireNumber = composeValidators(required, mustBeNumber);
 
-const AdSense = ({ member }) => (
+const DoubleClick = ({ member }) => (
   <>
-    <TextInput
-      name={`${member}.client`}
-      label="client"
-      placeholder="ca-pub-1234123412341234"
-      validate={required}
-    />
     <TextInput
       name={`${member}.slot`}
       label="slot"
-      placeholder="1234567890"
+      placeholder="/12345678/ad-roba_1"
       validate={required}
     />
     <TextInput
@@ -25,25 +20,20 @@ const AdSense = ({ member }) => (
       label="width"
       placeholder="300"
       validate={requireNumber}
-      parse={input => Number(input)}
+      parse={toNumber}
     />
     <TextInput
       name={`${member}.height`}
       label="height"
       placeholder="250"
       validate={requireNumber}
-      parse={input => Number(input)}
-    />
-    <TextInput
-      name={`${member}.format`}
-      label="format"
-      placeholder="rectangle"
+      parse={toNumber}
     />
   </>
 );
 
-AdSense.propTypes = {
+DoubleClick.propTypes = {
   member: PropTypes.string.isRequired,
 };
 
-export default AdSense;
+export default DoubleClick;
